use a small/dumb internal router since needs are simple
go-camo has simple needs, and a very simple dumb router is good enough.

This reduces dependencies, as well as keeps the actual proxy ServeHTTP
handler more generic -- and able to be plugged into larger apps if
desired.

In addition, this is an opportunity to move some things out of
camo.proxy into the dumb router layer.