Allow the HTTP server name to be dynamic
Updates the `go-camo` package to allow a user defined server name that
will replace the generic `go-camo` value upon being set. This can be
invoked using `--server-name`. For instance:

Included in this change is the ability to append the current version of
the package to the HTTP server value using `--expose-server-version`.
This assists in debugging and identifying individual versions when
running larger clusters of camo instances.