Ignore port when filtering IPs and Hostnames (#28)
The protection against proxying requests to local IPs can be bypassed by adding a port to the hostname (i.e. http://localhost:80)
This can result in a Server Side Request Forgery (SSRF) attack, allowing an attacker that controls the URL generated to impersonate the server in making requests.
This is fixed by using the Hostname() function from net/url, which only ever returns the hostname without the port.