proxy: Swap "content length exceeded" to be a HTTP 413
We've encountered a couple of cases where the content length of the
remote was over the threshold and while the error description is
helpful, the HTTP status code is not (currently a 404). Instead of
returning the "not found" status, we can update it to be a HTTP 413
which is "request entity too large". This will allow us to differentiate
resources that aren't really there and ones that are oversized.