Provide an explicit deny URL option
Camo currently ships with the option to provide a file path containing a
number of hostnames to `--allow-list` which limits where images will be
proxied from. This works great for restricting the source if you control
the origin or have a small number of origins.

What it doesn't cover very well is to explicitly deny access to a
resource on a fine grained level. Take for example, you use Camo to
proxy all third party assets to put a CDN in front of them and then one
of the origins used via Camo is hosting malicious or inappropriate
material. There currently isn't a sane way of preventing access to this
particular resource once anyone has the signed HMAC URL. There are work
arounds to kill the connection however they require drastic action like
rolling the HMAC which unless co-ordinated well, results in a service
interruption.

This change aims to address the above scenario by exposing a way to
define the inverse, an explicit deny list. It works on a similar
principle to the `allow-list` with one notable exception; it loosely
matches the outgoing URL. Instead of restricting based on hostname or
asset filename, it can match anywhere using `string.Contains`. This
brings with it great power and a potential to shoot yourself in the foot
if misconfigured. The thought behind this is that Re2 is pretty poor at
operations like negative lookaheads (specifically because it conflicts
with the O(n)-time guarantees of the library[1]) and I didn't want to
overload options with variants for hostname, path, etc. An added benefit
to this approach is that you're able to deny a single resource on a
domain and not the whole domain which is useful for shared image hosting
services.

[1]: https://groups.google.com/forum/#!topic/golang-nuts/7qgSDWPIh_E